twitter/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aniso8601=8.0.0=py_0
blas=1.0=mkl
ca-certificates=2020.10.14=0
certifi=2020.6.20=py38_0
chardet=4.0.0=pypi_0
click=7.1.2=pyhd3eb1b0_0
flask=1.1.2=pyhd3eb1b0_0
flask-restful=0.3.8=pyhd3eb1b0_0
flask-sqlalchemy=2.4.3=py_0
idna=2.10=pypi_0
intel-openmp=2020.2=254
itsdangerous=1.1.0=pyhd3eb1b0_0
jinja2=2.11.2=pyhd3eb1b0_0
markupsafe=1.1.1=py38he774522_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
numpy=1.20.1=pypi_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1k=h2bbff1b_0
pandas=1.2.3=pypi_0
pip=21.0.1=py38haa95532_0
pysocks=1.7.1=pypi_0
python=3.8.8=hdbf39b2_4
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2020.1=py_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
setuptools=52.0.0=py38haa95532_0
six=1.15.0=py38haa95532_0
sqlalchemy=1.3.18=pypi_0
sqlite=3.35.1=h2bbff1b_0
tweepy=3.10.0=pypi_0
tweet-preprocessor=0.6.0=pypi_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.4=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py38_0
